chore: refactor admin settings auth dialog into separate component
